add pathlib basic

+++ notebook/util_make_files.py
@@ -1,8 +1,10 @@
 import os
 import shutil
 
+
 def glob_example_detail():
-    shutil.rmtree('temp')
+    if os.path.exists('temp'):
+        shutil.rmtree('temp')
     
     os.makedirs('temp/dir/', exist_ok=True)
     os.makedirs('temp/dir/sub_dir1/', exist_ok=True)
@@ -21,3 +23,13 @@ def glob_example_detail():
     with open('temp/dir/sub_dir1/ccc.text', 'w'): pass
     
     with open('temp/dir/sub_dir2/ddd.text', 'w'): pass
+
+
+def pathlib_basic():
+    if os.path.exists('temp'):
+        shutil.rmtree('temp')
+    
+    os.makedirs('temp/dir/sub_dir/', exist_ok=True)
+
+    with open('temp/file.txt', 'w'): pass
+    with open('temp/dir/sub_dir/file2.txt', 'w'): pass
